requirements.txt
farm-haystack==1.14.0
jupyter==1.0.0
jupyterlab==3.6.1
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.1
sentence-transformers~=2.2.2
spacy==3.5.0
transformers>=4.25
torch>=1.12
